Fixed light bavure

+++ GolemEngine/Include/Components/Light/directional.h
@@ -8,10 +8,11 @@ class Shader;
 class GOLEM_ENGINE_API DirectionalLight : public Light
 {
 public:
-	int id;
+	size_t id;
 	Vector3 direction;
 
-	DirectionalLight(const Vector4& diffuse, const Vector4& ambient, const Vector4& specular, const Vector3& direction, const std::vector<DirectionalLight*>& _dirLights, int _maxDirs);
+	DirectionalLight(const Vector4& diffuse, const Vector4& ambient, const Vector4& specular, const Vector3& direction, 
+		const std::vector<DirectionalLight*>& _dirLights, size_t _maxDirs);
 	void SetDirectionalLight(Shader* shader);
 	
 };
